Removed old CountryTable component
